company-efficiency-optimizer/requirements_working.txt
# Working versions for Company Efficiency Optimizer
# Compatible with CrewAI 0.1.32 and Pydantic v2

# Core CrewAI and AI dependencies
crewai==0.1.32
langchain==0.1.0
langchain-core==0.1.23
langchain-community==0.0.20
langchain-openai==0.0.2.post1

# Pydantic v2 (required by CrewAI 0.1.32)
pydantic>=2.4.2,<3.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Vector database and embeddings
pinecone-client>=3.0.0
PyYAML>=6.0.0

# PDF processing and OCR
pytesseract>=0.3.10
pdf2image>=1.16.0,<3.0.0
Pillow>=10.0.0

# HTTP requests and API integration
requests>=2.31.0
Flask>=3.0.0

# Additional utilities
rich>=13.0.0
typer>=0.9.0

# OpenAI (required by CrewAI)
openai>=1.7.1,<2.0.0